fix: create new private paymaster client
